requirements.txt
slack_sdk==3.26.2
jira==3.6.0
requests==2.31.0
flask==3.0.1
flake8==7.0.0
slack_bolt==1.18.1
black==23.12.1
responses==0.24.1
python-dotenv==1.0.1
slack_sdk==3.26.2